final page first commit
This included some major changes. I have added a new floating images component that is re-used on landing and final page. I also changed a bit the animations for the opening page. we might discuss this. My solution on the final page is very wonky sometimes, because the setup is not ideal.